authorization

+++ app/Http/Controllers/DashboardController.php
@@ -77,7 +77,7 @@ class DashboardController extends Controller
     public function CreateAppForm() : View
     {
         $this->authorize('create', AppInfo::class);
-        
+
         $allAppInfos = app(AppStoreConnectController::class)->GetAppList()->getData();
         $allGitProjects = app(GithubController::class)->GetRepositories()->getData();
 
@@ -92,6 +92,8 @@ class DashboardController extends Controller
 
     public function StoreAppForm(StoreAppInfoRequest $request) : RedirectResponse
     {
+        $this->authorize('create', AppInfo::class);
+
         $createAppResponse = app(AppInfoController::class)->CreateApp($request)->getData();
         $projectName = $createAppResponse->app->project_name;
 
@@ -111,7 +113,12 @@ class DashboardController extends Controller
 
     public function SelectApp(GetAppInfoRequest $request) : View
     {
-        return view('appinfo-form')->with('appInfo', AppInfo::find($request->validated('id')));
+        // auth
+        $appId = $request->validated('id');
+        $app = AppInfo::find($appId);
+        $this->authorize('view', $app);
+
+        return view('appinfo-form')->with('appInfo', $app);
     }
 
     public function UpdateApp(UpdateAppInfoRequest $request): RedirectResponse
@@ -124,6 +131,11 @@ class DashboardController extends Controller
 
     public function DeleteApp(GetAppInfoRequest $request) : RedirectResponse
     {
+        // auth
+        $appId = $request->validated('id');
+        $app = AppInfo::find($appId);
+        $this->authorize('delete', $app);
+
         session()->flash('success', app(AppInfoController::class)->DeleteApp($request)->getData()->message);
 
         return to_route('get_app_list');
